pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "url_shortener"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
pydantic = "^1.10.7"
dependency-injector = "^4.41.0"
python-slugify = "^8.0.1"
flask-bootstrap = "^3.3.7.1"
qrcode = "^7.4.2"
python-environ = "^0.4.54"
flask-wtf = "^1.1.1"
gunicorn = "^20.1.0"
pymongo = {extras = ["srv"], version = "^4.3.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
